+++ packages/headless/src/controllers/core/did-you-mean/headless-core-did-you-mean.ts
@@ -6,6 +6,7 @@ import {configuration} from '../../../app/common-reducers';
 import {CoreEngine} from '../../../app/engine';
 import {
   applyDidYouMeanCorrection,
+  disableAutomaticQueryCorrection,
   enableDidYouMean,
 } from '../../../features/did-you-mean/did-you-mean-actions';
 import {didYouMeanReducer as didYouMean} from '../../../features/did-you-mean/did-you-mean-slice';
@@ -21,6 +22,16 @@ import {
 
 export type {QueryCorrection, WordCorrection};
 
+export interface DidYouMeanProps {
+  /**
+   * Whether to automatically apply corrections for queries that would otherwise return no results.
+   * When `automaticallyCorrectQuery` is `true`, the controller automatically triggers a new query using the suggested term.
+   * When `automaticallyCorrectQuery` is `false`, the controller returns the suggested term without triggering a new query.
+   *
+   * The default value is `true`.
+   */
+  automaticallyCorrectQuery?: boolean;
+}
 export interface DidYouMean extends Controller {
   /**
    * Apply query correction using the query correction, if any, currently present in the state.
@@ -66,8 +77,13 @@ export interface DidYouMeanState {
  * automatically triggers a new query with the suggested term.
  *
  * @param engine - The headless engine.
+ * @param props - The configurable `DidYouMean` controller properties.
+ * @returns A `DidYouMean` controller instance.
  */
-export function buildCoreDidYouMean(engine: CoreEngine): DidYouMean {
+export function buildCoreDidYouMean(
+  engine: CoreEngine,
+  props: DidYouMeanProps = {}
+): DidYouMean {
   if (!loadDidYouMeanReducers(engine)) {
     throw loadReducerError;
   }
@@ -76,6 +92,10 @@ export function buildCoreDidYouMean(engine: CoreEngine): DidYouMean {
   const {dispatch} = engine;
 
   dispatch(enableDidYouMean());
+
+  if (props.automaticallyCorrectQuery === false) {
+    dispatch(disableAutomaticQueryCorrection());
+  }
 
   const getState = () => engine.state;