+++ packages/headless/src/features/search-parameters/search-parameter-serializer.test.ts
@@ -51,6 +51,12 @@ describe('buildSearchParameterSerializer', () => {
       expect(result).toEqual('f-author=a,b&f-filetype=c,d');
     });
 
+    it('serializes the #fExcluded parameter correctly', () => {
+      const fExcluded = {author: ['a', 'b'], filetype: ['c', 'd']};
+      const result = serialize({fExcluded});
+      expect(result).toEqual('fExcluded-author=a,b&fExcluded-filetype=c,d');
+    });
+
     it('serializes special characters in the #f parameter correctly', () => {
       someSpecialCharactersThatNeedsEncoding.forEach((specialChar) => {
         const f = {author: ['a', specialChar]};
@@ -472,6 +478,7 @@ describe('buildSearchParameterSerializer', () => {
 
   it('can serialize and deserialize all search parameters', () => {
     const f = {author: ['a', 'b']};
+    const fExcluded = {source: ['a', 'b']};
     const cf = {geography: ['a', 'b']};
     const nf = {
       size: [buildNumericRange({start: 0, end: 10, state: 'selected'})],
@@ -487,7 +494,15 @@ describe('buildSearchParameterSerializer', () => {
     };
     const sf = {fileType: ['a', 'b']};
     const af = {documenttype: ['s', 'sd']};
-    const parameters = buildMockSearchParameters({f, cf, nf, df, sf, af});
+    const parameters = buildMockSearchParameters({
+      f,
+      cf,
+      nf,
+      df,
+      sf,
+      af,
+      fExcluded,
+    });
 
     const {serialize, deserialize} = buildSearchParameterSerializer();
     const serialized = serialize(parameters);